requirements.txt
torch>=2.0.0
torchaudio>=2.0.0
fastapi>=0.100.0
uvicorn>=0.23.0
google-generativeai>=0.3.0
pydantic>=2.0.0
pytest>=7.4.0
numpy>=1.24.0
tqdm>=4.65.0
